♻️ abstract requestEntries method
