+++ src/functions/searches/attach-to-one-explorer-search/attach-to-one-explorer-search.ts
@@ -6,40 +6,13 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { isBoolean, isEqual, isNil, isNull, uniqueId } from 'lodash';
-import {
-	BehaviorSubject,
-	combineLatest,
-	EMPTY,
-	firstValueFrom,
-	from,
-	lastValueFrom,
-	NEVER,
-	Observable,
-	of,
-	Subject,
-	Subscription,
-} from 'rxjs';
-import {
-	catchError,
-	concatMap,
-	distinctUntilChanged,
-	filter,
-	filter as rxjsFilter,
-	map,
-	shareReplay,
-	skipUntil,
-	startWith,
-	takeUntil,
-	tap,
-} from 'rxjs/operators';
+import { isBoolean, isEqual, isNull, uniqueId } from 'lodash';
+import { BehaviorSubject, combineLatest, EMPTY, from, lastValueFrom, Observable, Subject, Subscription } from 'rxjs';
+import { catchError, concatMap, distinctUntilChanged, filter, map, shareReplay, takeUntil, tap } from 'rxjs/operators';
 import {
 	ExplorerSearchEntries,
 	ExplorerSearchSubscription,
 	RawRequestSearchCloseMessageSent,
-	RawRequestSearchDetailsMessageSent,
-	RawRequestSearchStatsMessageSent,
-	RawRequestSearchStatsWithinRangeMessageSent,
 	RawResponseForSearchDetailsMessageReceived,
 	RawResponseForSearchStatsMessageReceived,
 	SearchEntries,
@@ -54,22 +27,19 @@ import { APIContext } from '../../utils';
 import { attachSearch } from '../attach-search';
 import { createRequiredSearchFilterObservable, DateRange } from '../helpers/create-required-search-filter-observable';
 import { makeSubscribeToOneRawSearch } from '../subscribe-to-one-raw-search';
-import { RawRequestExplorerSearchEntriesWithinRangeMessageSent } from '~/models';
 import {
 	countEntriesFromModules,
 	filterMessageByCommand,
 	getDefaultGranularityByRendererType,
-	recalculateZoomEnd,
-	RequiredSearchFilter,
 	SEARCH_FILTER_PREFIX,
 } from '../subscribe-to-one-search/helpers';
 import {
 	createInitialSearchFilter,
-	debouncedPooling,
 	extractZoomFromRawSearchStats,
 	mapToSearchStats,
 } from '~/functions/searches/helpers/attach-search';
 import { getPreviewDateRange, emitError } from '../helpers/attach-search';
+import { getRawRequestExplorerEntriesMsg, makeRequestEntries } from '~/functions/searches/helpers/request-entries';
 
 export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 	const subscribeToOneRawSearch = makeSubscribeToOneRawSearch(context);
@@ -215,164 +185,6 @@ export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 			takeUntil(close$),
 		);
 
-		const nextDetailsMsg = () =>
-			firstValueFrom(
-				searchMessages$.pipe(
-					filter(filterMessageByCommand(SearchMessageCommands.RequestDetails)),
-					// cleanup: Complete when/if the user calls .close()
-					takeUntil(close$),
-				),
-			);
-
-		let pollingSubs: Subscription;
-
-		const requestEntries = async (filter: RequiredSearchFilter): Promise<void> => {
-			if (closed) return undefined;
-
-			if (!isNil(pollingSubs)) {
-				pollingSubs.unsubscribe();
-			}
-			pollingSubs = new Subscription();
-
-			const filterID = uniqueId(SEARCH_FILTER_PREFIX);
-			filtersByID[filterID] = filter;
-
-			const first = filter.entriesOffset.index;
-			const last = first + filter.entriesOffset.count;
-			const startDate = filter.dateRange === 'preview' ? previewDateRange.start : filter.dateRange.start;
-			const start = startDate.toISOString();
-			const endDate = filter.dateRange === 'preview' ? previewDateRange.end : filter.dateRange.end;
-			const end = endDate.toISOString();
-			// TODO: Filter by .desiredGranularity and .fieldFilters
-
-			// Set up a promise to wait for the next details message
-			const detailsMsgP = nextDetailsMsg();
-			// Send a request for details
-			const requestDetailsMsg: RawRequestSearchDetailsMessageSent = {
-				type: searchTypeID,
-				data: { ID: SearchMessageCommands.RequestDetails, Addendum: { filterID } },
-			};
-			const detailsP = rawSubscription.send(requestDetailsMsg);
-
-			// Grab the results from the details response (we need it later)
-			const detailsResults = await Promise.all([detailsP, detailsMsgP]);
-			const detailsMsg = detailsResults[1];
-
-			// Keep sending requests for search details until Finished is true
-			pollingSubs.add(
-				rawSearchDetails$
-					.pipe(
-						// We've already received one details message - use it to start
-						startWith(detailsMsg),
-
-						// Extract the property that indicates if the data is finished
-						map(details => (details ? details.data.Finished : false)),
-
-						debouncedPooling({ rawSubscription, message: requestDetailsMsg }),
-
-						catchError(() => EMPTY),
-						takeUntil(close$),
-					)
-					.subscribe(),
-			);
-
-			const requestEntriesMsg: RawRequestExplorerSearchEntriesWithinRangeMessageSent = {
-				type: searchTypeID,
-				data: {
-					ID: SearchMessageCommands.RequestExplorerEntriesWithinRange,
-					Addendum: { filterID },
-					EntryRange: {
-						First: first,
-						Last: last,
-						StartTS: start,
-						EndTS: end,
-					},
-				},
-			};
-			// Keep sending requests for entries until finished is true
-			pollingSubs.add(
-				entries$
-					.pipe(
-						// Extract the property that indicates if the data is finished
-						map(entries => (entries ? entries.finished : false)),
-
-						debouncedPooling({ rawSubscription, message: requestEntriesMsg }),
-
-						catchError(() => EMPTY),
-						takeUntil(close$),
-					)
-					.subscribe(),
-			);
-			const entriesP = rawSubscription.send(requestEntriesMsg);
-
-			const requestStatsMessage: RawRequestSearchStatsMessageSent = {
-				type: searchTypeID,
-				data: {
-					ID: SearchMessageCommands.RequestAllStats,
-					Addendum: { filterID },
-					Stats: { SetCount: filter.overviewGranularity },
-				},
-			};
-			// Keep sending requests for stats until finished is true
-			pollingSubs.add(
-				rawSearchStats$
-					.pipe(
-						// Extract the property that indicates if the data is finished
-						map(stats => stats.data.Finished ?? false),
-
-						debouncedPooling({ rawSubscription, message: requestStatsMessage }),
-
-						catchError(() => EMPTY),
-						takeUntil(close$),
-					)
-					.subscribe(),
-			);
-			const statsP = rawSubscription.send(requestStatsMessage);
-
-			const requestStatsWithinRangeMsg: RawRequestSearchStatsWithinRangeMessageSent = {
-				type: searchTypeID,
-				data: {
-					ID: SearchMessageCommands.RequestStatsInRange,
-					Addendum: { filterID },
-					Stats: {
-						SetCount: filter.zoomGranularity,
-						SetEnd: recalculateZoomEnd(
-							detailsMsg ? detailsMsg.data.SearchInfo.MinZoomWindow : 1,
-							filter.zoomGranularity,
-							startDate,
-							endDate,
-						).toISOString(),
-						SetStart: start,
-					},
-				},
-			};
-			// Keep sending requests for stats-within-range until finished is true
-			pollingSubs.add(
-				rawStatsZoom$
-					.pipe(
-						// Extract the property that indicates if the data is finished
-						map(stats => stats.data.Finished ?? false),
-
-						debouncedPooling({ rawSubscription, message: requestStatsWithinRangeMsg }),
-
-						// Filter out finished events
-						shareReplay({ bufferSize: 1, refCount: true }),
-
-						catchError(() => EMPTY),
-						takeUntil(close$),
-					)
-					.subscribe(),
-			);
-			const statsRangeP = rawSubscription.send(requestStatsWithinRangeMsg);
-
-			await Promise.all([entriesP, statsP, detailsP, statsRangeP]);
-		};
-
-		filter$.subscribe(filter => {
-			requestEntries(filter);
-			setTimeout(() => requestEntries(filter), 2000); // TODO: Change this
-		});
-
 		const rawSearchStats$ = searchMessages$.pipe(
 			filter(filterMessageByCommand(SearchMessageCommands.RequestAllStats)),
 
@@ -393,6 +205,26 @@ export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 			// Complete when/if the user calls .close()
 			takeUntil(close$),
 		);
+
+		const requestEntries = makeRequestEntries({
+			close$,
+			entries$,
+			filtersByID,
+			previewDateRange,
+			rawSearchDetails$,
+			rawSearchStats$,
+			rawStatsZoom$,
+			rawSubscription,
+			searchMessages$,
+			searchTypeID,
+			getRawRequestEntriesMsg: getRawRequestExplorerEntriesMsg,
+			isClosed: () => closed,
+		});
+
+		filter$.subscribe(filter => {
+			requestEntries(filter);
+			setTimeout(() => requestEntries(filter), 2000); // TODO: Change this
+		});
 
 		const stats$ = combineLatest([
 			rawSearchStats$.pipe(distinctUntilChanged<RawResponseForSearchStatsMessageReceived>(isEqual)),
